requirements.txt
pywebview[cef]>=5.4
mss>=9.0.2
openai>=1.54.0
python-dotenv>=1.0.1
pyperclip>=1.9.0
Pillow>=10.4.0
pyautogui>=0.9.54
pywin32>=306
soundcard>=0.4.3
soundfile>=0.12.1
numpy>=1.26
imageio>=2.37
imageio-ffmpeg>=0.5
pynput>=1.7
